dir, refactor(feature/ujin/#85): api 폴더명 변경 및 api 요청 함수명 소문자로 변경
