+++ yeoreum/src/hooks/queries/notices.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
-import { requestGetNotices } from '../../api/notices';
+import { requestGetNotices } from '../../apis/notices';
 
 const useNoticesQuery = (token: string) => {
   // <데이터 타입, Error 타입, select 타입, query-key 타입>
