+++ yeoreum/src/pages/friend/index.tsx
@@ -3,7 +3,7 @@ import PostContainer from '../../components/board/PostContainer';
 import FriendTop from '../../components/friend/page/FriendTop';
 import UserSearch from '../../components/friend/page/UserSearch';
 import MyFriendList from '../../components/friend/page/MyFriendList';
-import { RequestGetSearchFriends } from '../../api/friends';
+import { requestGetSearchFriends } from '../../apis/friends';
 import { FriendResponseType } from '../../types/friend';
 
 function index() {
@@ -14,7 +14,7 @@ function index() {
   useEffect(() => {
     (async () => {
       setLoading(true);
-      const friends = await RequestGetSearchFriends(searchTerm);
+      const friends = await requestGetSearchFriends(searchTerm);
       setFriendList(friends);
 
       if (!searchTerm) {
